requirements.txt
Flask==2.2.2
requests==2.28.1
geopy==2.2.0
folium==0.14.0
python-dotenv-1.0.1